# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API fashion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/order-details/{id}": {
            "get": {
                "summary": "Show order details",
                "operationId": "6a2f1dd060dacb2eccd09cf49d3b9850",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "show order details.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            },
            "delete": {
                "summary": "Delete order",
                "operationId": "cda3f9927ff4e14ced8d63412aec46da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "order detail successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order-details": {
            "post": {
                "summary": "Create a new order and order details",
                "operationId": "294519256e6751e07791f359fd7efeb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string",
                                                    "example": "Description 1"
                                                },
                                                "price_unit": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "piece_id": {
                                                    "type": "string",
                                                    "example": "piece1"
                                                },
                                                "piece_type": {
                                                    "type": "string",
                                                    "example": "type1"
                                                },
                                                "piece_name": {
                                                    "type": "string",
                                                    "example": "Piece Name 1"
                                                },
                                                "piece_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "category_id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "category_name": {
                                                    "type": "string",
                                                    "example": "Category Name 1"
                                                },
                                                "texture_id": {
                                                    "type": "string",
                                                    "example": "texture1"
                                                },
                                                "texture_name": {
                                                    "type": "string",
                                                    "example": "Texture Name 1"
                                                },
                                                "texture_provider": {
                                                    "type": "string",
                                                    "example": "Provider 1"
                                                },
                                                "color_id": {
                                                    "type": "string",
                                                    "example": "color1"
                                                },
                                                "color_name": {
                                                    "type": "string",
                                                    "example": "Color Name 1"
                                                },
                                                "color_code": {
                                                    "type": "string",
                                                    "example": "Code1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "order": {
                                        "properties": {
                                            "neck": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 15.5
                                            },
                                            "shoulder": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 18
                                            },
                                            "arm": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 24
                                            },
                                            "mid_front": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 20
                                            },
                                            "bicep": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 12.5
                                            },
                                            "bust": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 36
                                            },
                                            "size": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10
                                            },
                                            "waist": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 30
                                            },
                                            "leg": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 32
                                            },
                                            "hip": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 40
                                            },
                                            "skirt_length": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 24
                                            },
                                            "unit_length": {
                                                "type": "string",
                                                "example": "inch"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "user": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "john.doe2@example.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "1234567890"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "USA"
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "New York"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order and order details created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "type": "object"
                                        },
                                        "order_details": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    }
}